+++ bundles/admin/admin-layereditor/view/Flyout.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { AdminLayerForm, AdminLayerFormHandler } from './AdminLayerForm';
-import { LayerWizard } from './LayerWizard';
+import { LayerWizard } from './LayerWizard/';
 import { Spin } from 'oskari-ui';
 import { LocaleProvider } from 'oskari-ui/util';
