+++ bundles/admin/admin-layereditor/view/LayerWizard/LayerWizard.stories.js
