+++ bundles/admin/admin-layereditor/view/LayerWizard/index.js
@@ -1,0 +1,1 @@
+export { LayerWizard } from './LayerWizard';
